Rename mikrotik coordinator module
