+++ homeassistant/components/mikrotik/__init__.py
@@ -7,8 +7,8 @@ from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
 from homeassistant.helpers import config_validation as cv, device_registry as dr
 
 from .const import ATTR_MANUFACTURER, DOMAIN
+from .coordinator import MikrotikDataUpdateCoordinator, get_api
 from .errors import CannotConnect, LoginError
-from .hub import MikrotikDataUpdateCoordinator, get_api
 
 CONFIG_SCHEMA = cv.removed(DOMAIN, raise_if_present=False)
