+++ homeassistant/components/mikrotik/coordinator.py
