+++ tests/components/mikrotik/test_device_tracker.py
@@ -82,7 +82,7 @@ async def test_device_trackers(
     device_2 = hass.states.get("device_tracker.device_2")
     assert device_2 is None
 
-    with patch.object(mikrotik.hub.MikrotikData, "command", new=mock_command):
+    with patch.object(mikrotik.coordinator.MikrotikData, "command", new=mock_command):
         # test device_2 is added after connecting to wireless network
         WIRELESS_DATA.append(DEVICE_2_WIRELESS)
 
@@ -150,7 +150,9 @@ async def test_arp_ping_success(
 ) -> None:
     """Test arp ping devices to confirm they are connected."""
 
-    with patch.object(mikrotik.hub.MikrotikData, "do_arp_ping", return_value=True):
+    with patch.object(
+        mikrotik.coordinator.MikrotikData, "do_arp_ping", return_value=True
+    ):
         await setup_mikrotik_entry(hass, arp_ping=True, force_dhcp=True)
 
         # test wired device_2 show as home if arp ping returns True
@@ -163,7 +165,9 @@ async def test_arp_ping_timeout(
     hass: HomeAssistant, mock_device_registry_devices
 ) -> None:
     """Test arp ping timeout so devices are shown away."""
-    with patch.object(mikrotik.hub.MikrotikData, "do_arp_ping", return_value=False):
+    with patch.object(
+        mikrotik.coordinator.MikrotikData, "do_arp_ping", return_value=False
+    ):
         await setup_mikrotik_entry(hass, arp_ping=True, force_dhcp=True)
 
         # test wired device_2 show as not_home if arp ping times out
@@ -262,7 +266,9 @@ async def test_update_failed(hass: HomeAssistant, mock_device_registry_devices) 
     await setup_mikrotik_entry(hass)
 
     with patch.object(
-        mikrotik.hub.MikrotikData, "command", side_effect=mikrotik.errors.CannotConnect
+        mikrotik.coordinator.MikrotikData,
+        "command",
+        side_effect=mikrotik.errors.CannotConnect,
     ):
         async_fire_time_changed(hass, utcnow() + timedelta(seconds=10))
         await hass.async_block_till_done(wait_background_tasks=True)
